+++ indico/core/signals/plugin.py
@@ -56,11 +56,19 @@ definitions.
 ''')
 
 get_conference_themes = _signals.signal('get-conference-themes', '''
-Expected to return ``(name, css, title)`` tuples for conference stylesheets.
-``name`` is the internal name used for the stylesheet which will be
-stored when the theme is selected in an event.  ``css`` is the location
-of the CSS file, relative to the plugin's ``static`` folder.  ``title``
-is the title displayed to the user when selecting the theme.
+Expected to return one or more :class:`indico.modules.events.layout.util.ConferenceTheme`
+objects containing the information about custom conference themes.
+
+Required:
+- ``name``     -- string indicating the internal name used for the stylesheet which will be
+                  stored when the theme is selected in an event.
+- ``css_path`` -- string indicating the location of the CSS file, relative to the
+                  plugin's ``static`` folder.
+- ``title``    -- string indicating the title displayed to the user when selecting the theme.
+
+Optional:
+- ``js_path``  -- string indicating the location for a simple, static javascript file to be
+                  loaded, relative to the plugin's ``static`` folder.
 ''')
 
 get_template_customization_paths = _signals.signal('get-template-customization-paths', '''