+++ indico/modules/core/controllers.py
@@ -26,6 +26,7 @@ from indico.core.sentry import submit_user_feedback
 from indico.core.settings import PrefixSettingsProxy
 from indico.modules.admin import RHAdminBase
 from indico.modules.cephalopod import cephalopod_settings
+from indico.modules.core.captcha import generate_captcha_challenge, get_captcha_plugin
 from indico.modules.core.forms import SettingsForm
 from indico.modules.core.settings import core_settings, social_settings
 from indico.modules.core.views import WPContact, WPSettings
@@ -325,3 +326,29 @@ class RHConfig(RH):
                        has_privacy_policy=bool(privacy_policy_url or privacy_policy_html),
                        privacy_policy_html=privacy_policy_html,
                        privacy_policy_url=privacy_policy_url)
+
+
+class RHAPIGenerateCaptcha(RH):
+    """Generate a CAPTCHA.
+
+    If a CAPTCHA plugin is available it is used,
+    otherwise we fallback to the built-in CAPTCHA.
+
+    See `CaptchaPluginMixin` for more information about
+    custom CAPTCHA plugins.
+    """
+
+    def _process_GET(self):
+        if plugin := get_captcha_plugin():
+            data = plugin.generate_captcha()
+        else:
+            data, answer = generate_captcha_challenge()
+            session['captcha_state'] = answer
+        rv = jsonify(data)
+        # make sure browsers don't cache this. otherwise using the back button after successfully
+        # using the old captcha code will go back to the page showing the already-used code instead
+        # of fetching a new one
+        rv.headers['Cache-Control'] = 'no-cache, no-store, must-revalidate, max-age=0'
+        rv.headers['Pragma'] = 'no-cache'
+        rv.headers['Expires'] = '0'
+        return rv