+++ indico/modules/events/layout/util.py
@@ -6,6 +6,7 @@
 # LICENSE file for more details.
 
 from collections import defaultdict
+from dataclasses import dataclass
 from itertools import chain, count
 
 from sqlalchemy.exc import IntegrityError
@@ -308,19 +309,43 @@ def is_menu_entry_enabled(entry_name, event):
     return get_menu_entry_by_name(entry_name, event).is_enabled
 
 
+@dataclass(frozen=True)
+class ConferenceTheme:
+    """
+    Holds the values of a given conference theme to be used by Indico.
+
+    Required:
+    - ``name``     -- string indicating the internal name used for the stylesheet which will be
+                      stored when the theme is selected in an event.
+    - ``css_path`` -- string indicating the relative location of the CSS file
+    - ``title``    -- string indicating the title displayed to the user when selecting the theme.
+
+    Optional:
+    - ``js_path``  -- string indicating the relative location for a simple, static javascript file
+    """
+
+    name: str
+    css_path: str
+    title: str
+    js_path: str = None
+
+
 def get_plugin_conference_themes():
     data = values_from_signal(signals.plugin.get_conference_themes.send(), return_plugins=True)
-    return {':'.join((plugin.name, name)): (path, title) for plugin, (name, path, title) in data}
+    # backwards compatibility in case theme_info is a tuple instead of a `ConferenceTheme``
+    data = [(plugin, (theme_info if isinstance(theme_info, ConferenceTheme) else ConferenceTheme(*theme_info)))
+            for plugin, theme_info in data]
+    return {f'{plugin.name}:{theme_info.name}': theme_info for plugin, theme_info in data}
 
 
 def _build_css_url(theme):
     if ':' in theme:
         try:
-            path = get_plugin_conference_themes()[theme][0]
+            path = get_plugin_conference_themes()[theme].css_path
         except KeyError:
             return None
         plugin = theme.split(':', 1)[0]
-        return url_for_plugin(plugin + '.static', filename=path)
+        return url_for_plugin(f'{plugin}.static', filename=path)
     else:
         css_base = url_parse(config.CONFERENCE_CSS_TEMPLATES_BASE_URL).path
         return f'{css_base}/{theme}'
@@ -347,6 +372,39 @@ def get_css_url(event, force_theme=None, for_preview=False):
         return url_for('event_layout.css_display', event, slug=event.stylesheet_metadata['hash'])
     elif layout_settings.get(event, 'theme'):
         return _build_css_url(layout_settings.get(event, 'theme'))
+
+
+def _build_js_url(theme):
+    if ':' not in theme:
+        return None
+    try:
+        if not (path := get_plugin_conference_themes()[theme].js_path):
+            return None
+    except KeyError:
+        return None
+    plugin = theme.split(':', 1)[0]
+    return url_for_plugin(f'{plugin}.static', filename=path)
+
+
+def get_js_url(event, force_theme=None, for_preview=False):
+    """Build the URL of a JS resource.
+
+    :param event: The `Event` to get the JS url for
+    :param force_theme: The ID of the theme to import the custom JS resource
+                        only if it exists
+    :param for_preview: Whether the URL is used in the JS preview page
+    :return: The URL to the JS resource
+    """
+    from indico.modules.events.layout import layout_settings
+
+    if force_theme and force_theme != '_custom':
+        return _build_js_url(force_theme)
+    elif for_preview and force_theme is None:
+        return None
+    elif force_theme == '_custom' or layout_settings.get(event, 'use_custom_css'):
+        return None
+    elif layout_settings.get(event, 'theme'):
+        return _build_js_url(layout_settings.get(event, 'theme'))
 
 
 def get_logo_data(event):