+++ indico/web/client/js/react/components/index.js
@@ -55,3 +55,4 @@ export {default as WTFRemoteSearchDropdown} from './WTFRemoteSearchDropdown';
 export {default as ICSCalendarLink} from './ICSCalendarLink';
 export {default as Favorite} from './Favorite';
 export {default as PersonDetailsModal} from './PersonDetailsModal';
+export {default as Captcha} from './Captcha';